+++ R/apportion.R
@@ -17,7 +17,8 @@
 #' date_admitted = lubridate::dmy("01-01-2015"), specimen_date =
 #' lubridate::dmy("05-01-2015"), date_entered = lubridate::dmy("26-10-2015"))
 #' \dontrun{
-#' apportion("mssa", "NHS Acute Trust", "In-patient", date_admitted = NA,
+#' apportion("mssa", "NHS Acute Trust", "In-patient",
+#' date_admitted = as.Date(NA_real_, origin = "01-01-1970"),
 #' specimen_date = lubridate::dmy("01-01-2015"), date_entered =
 #' lubridate::dmy("26-10-2015"))
 #' apportion("mssa", "NHS Acute Trust", "In-patient", date_admitted =