+++ R/zzz.R
@@ -2,6 +2,6 @@
   packageStartupMessage(
     paste0(
       "\n\r\n\rThanks for using the HCAIDCS package.",
-      "\n\r\n\rPlease note that the apportion_prior_healthcare function is still experimental and has not been validated by a second team member or against the DCS.")
+      "\n\r\n\rIf you have any suggestions to improve the HCAIDCS package, please file an issue at https://github.com/PublicHealthEngland/hcaidcs/issues")
   )
 }